implement command line parsing
I added a basic command line parsing system which can detect which
subcommand was used and run a function based on it. I also added a nice
help command which is also run on most errors. Finally I added the
earlier api parsing work to the add command so it can be tested with
something like `modget add 394468`.