add: basic loader filtering
I added a list of valid loaders and a loader validation function.
Additionally I added a basic function to filter the files for ones that
contain the specified loader. I still need to handle having NO loader as
input and having no loaders listed as a GameVersion (most mods that only
support one loader will do this so we must just assume it's correct...)