add: loader filtering edge cases
I added a dedicated type StringSet and improved the loaderFilter
function to add the file if it doesn't specify any conflicting loaders.
This is mostly important for 1.12 and below where adding loader tags was
not common. Additionally I made it so the loader and version warnings
will not display if none are passed in.